PR Feedback. Refactor governance math into different file, usage of tenary operator, and post duration change hook
