+++ governance/src/components/Profile/StakingSummary.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { PrimaryText, SecondaryText, Title } from "shared/lib/designSystem";
 import theme from "shared/lib/designSystem/theme";
 import colors from "shared/lib/designSystem/colors";
-import { calculateInitialveRBNAmount } from "shared/lib/utils/math";
+import { calculateInitialveRBNAmount } from "shared/lib/utils/governanceMath";
 import TooltipExplanation from "shared/lib/components/Common/TooltipExplanation";
 import HelpInfo from "shared/lib/components/Common/HelpInfo";
 import { useRBNTokenAccount } from "shared/lib/hooks/useRBNTokenSubgraph";
