+++ governance/src/components/Staking/StakingModalUpdatePreview.tsx
@@ -13,15 +13,12 @@ import {
 } from "shared/lib/designSystem";
 import colors from "shared/lib/designSystem/colors";
 import { StakeIcon } from "shared/lib/assets/icons/icons";
-import {
-  formatBigNumber,
-  formatBigNumberAmount,
-  calculateInitialveRBNAmount,
-} from "shared/lib/utils/math";
+import { formatBigNumber, formatBigNumberAmount } from "shared/lib/utils/math";
 import { ActionButton } from "shared/lib/components/Common/buttons";
 import theme from "shared/lib/designSystem/theme";
 import { StakingUpdateMode } from "./types";
 import { useRBNTokenAccount } from "shared/lib/hooks/useRBNTokenSubgraph";
+import { calculateInitialveRBNAmount } from "shared/lib/utils/governanceMath";
 
 const ModalBackButton = styled.div`
   display: flex;
